+++ Tests/SwiftLintFrameworkTests/VerticalWhitespaceRuleTests.swift
@@ -11,6 +11,8 @@ import XCTest
 
 class VerticalWhitespaceRuleTests: XCTestCase {
 
+    private let ruleID = VerticalWhitespaceRule.description.identifier
+
     func testVerticalWhitespaceWithDefaultConfiguration() {
         // Test with default parameters
         verifyRule(VerticalWhitespaceRule.description)
@@ -25,5 +27,43 @@ class VerticalWhitespaceRuleTests: XCTestCase {
 
         verifyRule(maxEmptyLinesDescription,
                    ruleConfiguration: ["max_empty_lines": 2])
+    }
+
+    func testAutoCorrection() {
+        let maxEmptyLinesDescription = VerticalWhitespaceRule.description
+            .with(nonTriggeringExamples: [])
+            .with(triggeringExamples: [])
+            .with(corrections: [
+                "let b = 0\n\n↓\n↓\n↓\n\nclass AAA {}\n": "let b = 0\n\n\nclass AAA {}\n",
+                "let b = 0\n\n\nclass AAA {}\n": "let b = 0\n\n\nclass AAA {}\n"
+                ])
+
+        verifyRule(maxEmptyLinesDescription,
+                   ruleConfiguration: ["max_empty_lines": 2])
+    }
+
+    func testViolationMessageWithMaxEmptyLines() {
+        guard let config = makeConfig(["max_empty_lines": 2], ruleID) else {
+            XCTFail("Failed to create configuration")
+            return
+        }
+        let allViolations = violations("let aaaa = 0\n\n\n\nlet bbb = 2\n", config: config)
+
+        let verticalWhiteSpaceViolation = allViolations.first { $0.ruleDescription.identifier == ruleID }
+        if let violation = verticalWhiteSpaceViolation {
+            XCTAssertEqual(violation.reason, "Limit vertical whitespace to maximum 2 empty lines. Currently 3.")
+        } else {
+            XCTFail("A vertical white space violation should have been triggered!")
+        }
+    }
+
+    func testViolationMessageWithDefaultConfiguration() {
+        let allViolations = violations("let aaaa = 0\n\n\n\nlet bbb = 2\n")
+        let verticalWhiteSpaceViolation = allViolations.first(where: { $0.ruleDescription.identifier == ruleID })
+        if let violation = verticalWhiteSpaceViolation {
+            XCTAssertEqual(violation.reason, "Limit vertical whitespace to a single empty line. Currently 3.")
+        } else {
+            XCTFail("A vertical white space violation should have been triggered!")
+        }
     }
 }